# [Enhancement] Optimize parquet scanner by arrow io coalesce