--- a/be/src/common/config.h
+++ b/be/src/common/config.h
@@ -1256,4 +1256,8 @@ CONF_mInt32(olap_string_max_length, "1048576");
 // if mem_limit < 16 GB, disable JIT.
 // else it = min(mem_limit*0.01, 1GB)
 CONF_mInt64(jit_lru_cache_size, "0");
+
+CONF_mInt64(arrow_io_coalesce_read_max_buffer_size, "8388608");
+CONF_mInt64(arrow_io_coalesce_read_max_distance_size, "1048576");
+CONF_mInt64(arrow_read_batch_size, "4096");
 } // namespace starrocks::config