--- a/be/src/exec/parquet_reader.cpp
+++ b/be/src/exec/parquet_reader.cpp
@@ -47,8 +47,6 @@ ParquetReaderWrap::ParquetReaderWrap(std::shared_ptr<arrow::io::RandomAccessFile
           _read_size(read_size) {
     _parquet = std::move(parquet_file);
     _properties = parquet::ReaderProperties();
-    _properties.enable_buffered_stream();
-    _properties.set_buffer_size(1 * 1024 * 1024);
     _filename = (reinterpret_cast<ParquetChunkFile*>(_parquet.get()))->filename();
 }
 
@@ -84,6 +82,28 @@ Status ParquetReaderWrap::_init_parquet_reader() {
         * A DATETIME or TIMESTAMP value can include a trailing fractional seconds part in up to microseconds (6 digits) precision
         */
         arrow_reader_properties.set_coerce_int96_timestamp_unit(arrow::TimeUnit::MICRO);
+
+        // default batch size is 64K
+        arrow_reader_properties.set_batch_size(config::arrow_read_batch_size);
+
+        // io coalesce
+        // performance 0: tpcds store_sales, 23 columns, 649M, 7218819 lines
+        //                  | time | file read count | memory
+        // io coalesce 8M   | 13s  |   80            | 587M
+        // buffer stream 8M | 15s  |   176           | 1313M
+        // buffer stream 1M | 29s  |   1145          | 1157M
+        //
+        // performance 1: 1001 columns table, 147M, 50000 lines
+        //                  | time | file read count | memory
+        // io coalesce 8M   | 3s   |   20            | 1G
+        // buffer stream 8M | 15s  |   1003          | 3.3G
+        // buffer stream 1M | 15s  |   1003          | 10.1G
+        //
+        arrow_reader_properties.set_pre_buffer(true);
+        auto cache_options = arrow::io::CacheOptions::LazyDefaults();
+        cache_options.hole_size_limit = config::arrow_io_coalesce_read_max_distance_size;
+        cache_options.range_size_limit = config::arrow_io_coalesce_read_max_buffer_size;
+        arrow_reader_properties.set_cache_options(cache_options);
 
         // new file reader for parquet file
         auto st = parquet::arrow::FileReader::Make(arrow::default_memory_pool(),